+++ wandb/sdk/artifacts/storage_policies/wandb_storage_policy.py
@@ -403,7 +403,8 @@ class WandbStoragePolicy(StoragePolicy):
                 with cache_open("wb") as f, open(entry.local_path, "rb") as src:
                     shutil.copyfileobj(src, f)
                 if entry.local_path.startswith(staging_dir):
-                    # Move staged files to cache instead
+                    # Delete staged files as soon as they're copied to the cache
+                    # instead of waiting till all the files are uploaded
                     os.remove(entry.local_path)
             except OSError as e:
                 termwarn(f"Failed to cache {entry.local_path}, ignoring {e}")